Component boot removal
The modifications to boot were made under the assumption that StorageOS would not be used on a device with a screen. Seeing as this may no longer be the plan and considering the file is only loaded into memory while booting, these changes are no longer neccessary. So it has been added to the files pulled from openOS at installation.